pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kernelmoments"
version = "0.1.0"
description = "Data visualisation and estimation tools based on kernel regression"
readme = "README_PYPI.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "tlaiho", email = "tslaiho@gmail.com"},
]
keywords = ["kernel-regression", "conditional-moments", "nonparametric", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "kernelboost>=0.3.1",
    "numpy",
    "matplotlib",
]

[project.urls]
Repository = "https://github.com/tlaiho/kernelmoments"
Issues = "https://github.com/tlaiho/kernelmoments/issues"

[project.optional-dependencies]
gpu = ["cupy-cuda12x"]

[tool.pytest.ini_options]
python_files = ["*_tests.py", "test_*.py"]
